DELIA-41091: Video freeze when player transition from ads to content
Reason for change: Avoid sending flush start/stop events right after
a flushing seek which pauses basesrc src_loop task and leads to freeze
Test Procedure: tested on pace xi5
Risks: Low

+++ priv_aamp.h
@@ -3058,7 +3058,7 @@ public:
 	 *   @param[in] rate - playback rate
 	 *   @return void
 	 */
-	void FlushStreamSink(double position, int rate);
+	void FlushStreamSink(double position, double rate);
 
 private:
